Active state for Nav
Hover and active state has coded, index JS has been added in Pages and components for export and import.
