+++ src/components/header/Header.css
@@ -13,6 +13,7 @@
   display: flex;
   align-items: center;
   width: 50%;
+  margin: 0 1rem 0 1rem;
   border: 1px solid var(--gray100);
   border-radius: var(--radius);
 }
