+++ src/components/header/Header.jsx
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import logo from "../../images/space.png";
 import { FiSearch } from "react-icons/fi";
-import Button from "../button/Button";
-import Modal from "../modal/Modal";
+import Button from "../Button/Button";
+import Modal from "../Modal/Modal";
 import "./Header.css";
 
 const Header = () => {
