+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Button } from "./Button/Button";
+export { default as Header } from "./Header/Header";
+export { default as Modal } from "./Modal/Modal";
+export { default as NavBar } from "./NavBar/NavBar";
