+++ src/components/modal/Modal.css
@@ -8,7 +8,7 @@
   background-color: hsla(0, 0%, 0%, 0.7);
   display: grid;
   place-items: center;
-  z-index: 1;
+  z-index: 9;
 }
 /* modal content styles */
 .modal__content {
@@ -36,6 +36,11 @@
   font-size: var(--fs-big);
   color: var(--gray700);
   cursor: pointer;
+}
+
+.close-btn:hover {
+  background-color: var(--gray100);
+  border-radius: 50%;
 }
 
 .modal__content > form {
