+++ src/components/navbar/NavBar.css
@@ -1,35 +1,28 @@
-.navBar {
-  border-top: 1px solid var(--gray100);
-  border-bottom: 1px solid var(--gray100);
+nav {
+  width: 100%;
+  border: 1px solid rgba(100, 100, 100, 0.2);
+  top: 80px;
 }
 
-.navBar__items {
-  font-size: var(--fs-small);
+.navBar__links {
+  color: var(--gray700);
   font-weight: var(--fw-medium);
   display: flex;
   justify-content: space-evenly;
+  padding: 0.8rem;
 }
 
-.navBar__item {
-  position: relative;
-  padding: 0.8rem 0;
+.navBar__links a {
+  transition: all 0.3s ease-in-out;
+  padding: 0.5rem;
 }
 
-.navBar__item::before {
-  content: "";
-  position: absolute;
-  display: block;
-  height: 4px;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  border-radius: var(--radius);
-  background-color: var(--primary);
-  opacity: 0;
-  transition: 0.5s ease-in-out;
+.navBar__links a:hover {
+  background-color: var(--gray100);
+  border-radius: 1.5rem;
 }
 
-.navBar__item:hover::before,
-.navBar__item.active::before {
-  opacity: 1;
+.active {
+  background-color: var(--gray100);
+  border-radius: 1.5rem;
 }
