+++ src/index.css
@@ -66,7 +66,10 @@ svg {
 }
 
 /* anchor tags styling */
+ul,
+li,
 a {
+  list-style: none;
   text-decoration: none;
   color: var(--primary-dark);
 }
