--- src/pages/Fashion.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Fashion = () => {
-    return (
-        <div className='Fashion container'>Fashion</div>
-    )
-}
-
-export default Fashion
