+++ src/pages/Fashion/Fashion.css
@@ -1,0 +1,1 @@
+
