+++ src/pages/Fashion/Fashion.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Fashion.css";
+
+const Fashion = () => {
+  return <div className="Fashion container">Fashion</div>;
+};
+
+export default Fashion;
