--- src/pages/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Home = () => {
-    return (
-        <div className='Home container'>Home</div>
-    )
-}
-
-export default Home
