+++ src/pages/Home/Home.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Home.css";
+
+const Home = () => {
+  return <div className="Home container">Home</div>;
+};
+
+export default Home;
