--- src/pages/Jewels.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Jewels = () => {
-    return (
-        <div className='Jewels container'>Jewels</div>
-    )
-}
-
-export default Jewels
