+++ src/pages/Jewels/Jewels.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Jewels.css";
+
+const Jewels = () => {
+  return <div className="Jewels container">Jewels</div>;
+};
+
+export default Jewels;
