+++ src/pages/Makeup/Makeup.jsx
@@ -1,5 +1,7 @@
 import React from "react";
 
+import "./Makeup.css";
+
 const Makeup = () => {
   return <div className="Makeup container">MakeUp</div>;
 };
