--- src/pages/Models.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Models = () => {
-    return (
-        <div className='Models container'>Models</div>
-    )
-}
-
-export default Models
