+++ src/pages/Models/Models.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Models.css";
+
+const Models = () => {
+  return <div className="Models container">Models</div>;
+};
+
+export default Models;
