--- src/pages/Photo.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Photo = () => {
-    return (
-        <div className='Photo container'>Photo</div>
-    )
-}
-
-export default Photo
