+++ src/pages/Photo/Photo.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Photo.css";
+
+const Photo = () => {
+  return <div className="Photo container">Photo</div>;
+};
+
+export default Photo;
