--- src/pages/Rent.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Rent = () => {
-    return (
-        <div className='Rent container'>Rent</div>
-    )
-}
-
-export default Rent
