+++ src/pages/Rent/Rent.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Rent.css";
+
+const Rent = () => {
+  return <div className="Rent container">Rent</div>;
+};
+
+export default Rent;
