--- src/pages/Studio.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Studio = () => {
-    return (
-        <div className='Studio container'>Studio</div>
-    )
-}
-
-export default Studio
