+++ src/pages/Studio/Studio.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Studio.css";
+
+const Studio = () => {
+  return <div className="Studio container">Studio</div>;
+};
+
+export default Studio;
