--- src/pages/Video.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Video = () => {
-    return (
-        <div className='Video container'>Video</div>
-    )
-}
-
-export default Video
