+++ src/pages/Video/Video.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./Video.css";
+
+const Video = () => {
+  return <div className="Video container">Video</div>;
+};
+
+export default Video;
