+++ src/pages/index.js
@@ -1,0 +1,9 @@
+export { default as Fashion } from "./Fashion/Fashion";
+export { default as Home } from "./Home/Home";
+export { default as Jewels } from "./Jewels/Jewels";
+export { default as Makeup } from "./Makeup/Makeup";
+export { default as Models } from "./Models/Models";
+export { default as Photo } from "./Photo/Photo";
+export { default as Rent } from "./Rent/Rent";
+export { default as Studio } from "./Studio/Studio";
+export { default as Video } from "./Video/Video";
